Remove timeable concern
This was very specific to the match model and was going to be hard to
reuse on other models. Removing and merging into the Match class.
